Changed the lib and utils folder location
